requirements.txt
opencv-python-headless
tesserocr
streamlit
reportlab
Pillow